ubuntu/icewm container compat w/ Ubuntu 20.04

+++ src/ubuntu/install/tools.sh
@@ -8,7 +8,9 @@ apt-get install -y vim wget net-tools locales bzip2 \
     python-numpy #used for websockify/novnc
 apt-get clean -y
 
-ln -s /usr/bin/python2 /usr/bin/python
+if [[ ! -e /usr/bin/python ]]; then
+    ln -s /usr/bin/python2 /usr/bin/python
+fi
 
 echo "generate locales für en_US.UTF-8"
 locale-gen en_US.UTF-8
